hv_func.h: remove stadtx and zaphod hash
I would prefer not to carry the weight of the Perl hash function
on my shoulder. It is too much responsibility, and there are other
hash functions available that are known to more secure and still
relatively fast. So we shall use them instead.

The new default for 64 bit builds will be Siphash 1-3 (SIPHASH13),
for 32 bit builds it will be Chaskey. https://mouha.be/chaskey/

We should document what to do to add a new hash function so people
can easily use their own.

One problem we have is that since the hash function is a macro
XS modules must be recompiled if we change the hash function and
they use any hashing functions. :-( I really wish we could change
the hash function and not have to do that.